Refactor JSON tests to make them reusable in jackson v3
