+++ vertx-core/src/test/java/io/vertx/tests/json/jackson/JacksonDatabindTest.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.tests.json;
+package io.vertx.tests.json.jackson;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
